+++ src/integrationCapture.ts
@@ -131,6 +131,10 @@ const integrationMappingExternal: IntegrationIdMapping = {
     },
     _epik: {
         mappedKey: 'Pinterest.click_id',
+    // Snapchat
+    // https://developers.snap.com/api/marketing-api/Conversions-API/UsingTheAPI#sending-click-id
+    _scid: {
+        mappedKey: 'SnapchatConversions.Cookie1',
         output: IntegrationOutputs.CUSTOM_FLAGS,
     },
 };