+++ src/stub/mparticle.stub.js
@@ -166,6 +166,7 @@ function createAliasRequest() {
         destinationMpid: 'b',
         startTime: new Date().getTime(),
         endTime: new Date().getTime(),
+        scope: 'device',
     };
 }