+++ test/src/tests-native-sdk.js
@@ -7,6 +7,32 @@ const getLocalStorage = Utils.getLocalStorage,
     mParticleAndroid = Utils.mParticleAndroid,
     HTTPCodes = Constants.HTTPCodes;
 
+// Alias request test fixtures
+const aliasRequestFixture = {
+    destinationMpid: '101',
+    sourceMpid: '202',
+    startTime: 300,
+    endTime: 400,
+    scope: 'device',
+};
+
+const expectedAliasPayload = {
+    DestinationMpid: '101',
+    SourceMpid: '202',
+    StartUnixtimeMs: 300,
+    EndUnixtimeMs: 400,
+};
+
+const expectedAliasPayloadWithDeviceScope = JSON.stringify({
+    ...expectedAliasPayload,
+    Scope: 'device',
+});
+
+const expectedAliasPayloadWithMpidScope = JSON.stringify({
+    ...expectedAliasPayload,
+    Scope: 'mpid',
+});
+
 describe('native-sdk methods', function() {
     beforeEach(function() {
         mParticle._resetForTests(MPConfig);
@@ -649,20 +675,36 @@ describe('native-sdk methods', function() {
                 mParticleAndroidV2Bridge.modifyData.should.equal(JSONData);
             });
 
-            it('should send a JSON object to the Android\'s Alias method', () => {
+            it("should send a JSON object to the Android's Alias method", () => {
                 let callbackResult;
-                const aliasRequest = {
-                    destinationMpid: '101',
-                    sourceMpid: '202',
-                    startTime: 300,
-                    endTime: 400,
-                };
 
-                mParticle.Identity.aliasUsers(aliasRequest, function(callback) {
+                mParticle.Identity.aliasUsers(aliasRequestFixture, function(callback) {
                     callbackResult = callback;
                 });
                 mParticleAndroidV2Bridge.aliasUsers.should.equal(
-                    '{"DestinationMpid":"101","SourceMpid":"202","StartUnixtimeMs":300,"EndUnixtimeMs":400}'
+                    expectedAliasPayloadWithDeviceScope
+                );
+
+                callbackResult.httpCode.should.equal(
+                    HTTPCodes.nativeIdentityRequest
+                );
+                callbackResult.message.should.equal(
+                    'Alias request sent to native sdk'
+                );
+            });
+
+            it("should send a JSON object with scope to the Android's Alias method when scope is provided", () => {
+                let callbackResult;
+                const aliasRequestWithMPIDScope = {
+                    ...aliasRequestFixture,
+                    scope: 'mpid',
+                };
+
+                mParticle.Identity.aliasUsers(aliasRequestWithMPIDScope, function(callback) {
+                    callbackResult = callback;
+                });
+                mParticleAndroidV2Bridge.aliasUsers.should.equal(
+                    expectedAliasPayloadWithMpidScope
                 );
 
                 callbackResult.httpCode.should.equal(
@@ -960,16 +1002,10 @@ describe('native-sdk methods', function() {
                 mParticleIOSV2Bridge.reset();
             });
 
-            it('should send a JSON object to the iOS SDK\'s Alias method', () => {
+            it("should send a JSON object to the iOS SDK's Alias method", () => {
                 let callbackResult;
-                const aliasRequest = {
-                    destinationMpid: '101',
-                    sourceMpid: '202',
-                    startTime: 300,
-                    endTime: 400,
-                };
 
-                mParticle.Identity.aliasUsers(aliasRequest, function(callback) {
+                mParticle.Identity.aliasUsers(aliasRequestFixture, function(callback) {
                     callbackResult = callback;
                 });
 
@@ -981,9 +1017,37 @@ describe('native-sdk methods', function() {
                 );
                 JSON.stringify(
                     JSON.parse(mParticleIOSV2Bridge.data[0]).value
-                ).should.equal(
-                    '{"DestinationMpid":"101","SourceMpid":"202","StartUnixtimeMs":300,"EndUnixtimeMs":400}'
+                ).should.equal(expectedAliasPayloadWithDeviceScope);
+                mParticleIOSV2Bridge.reset();
+
+                callbackResult.httpCode.should.equal(
+                    HTTPCodes.nativeIdentityRequest
                 );
+                callbackResult.message.should.equal(
+                    'Alias request sent to native sdk'
+                );
+            });
+
+            it("should send a JSON object with scope to the iOS SDK's Alias method when scope is provided", () => {
+                let callbackResult;
+                const aliasRequestWithMPIDScope = {
+                    ...aliasRequestFixture,
+                    scope: 'mpid',
+                };
+
+                mParticle.Identity.aliasUsers(aliasRequestWithMPIDScope, function(callback) {
+                    callbackResult = callback;
+                });
+
+                JSON.parse(mParticleIOSV2Bridge.data[0]).should.have.properties(
+                    ['path', 'value']
+                );
+                JSON.parse(mParticleIOSV2Bridge.data[0]).path.should.equal(
+                    'aliasUsers'
+                );
+                JSON.stringify(
+                    JSON.parse(mParticleIOSV2Bridge.data[0]).value
+                ).should.equal(expectedAliasPayloadWithMpidScope);
                 mParticleIOSV2Bridge.reset();
 
                 callbackResult.httpCode.should.equal(